refactor: extract useQuizProgress hook and remove thin wrapper question files (#116)
- Add `hooks/quiz/useQuizProgress` — consolidates the 6-line
  quizGameStore selector block that was duplicated in every custom
  question component into a single hook call
- Apply hook to ClockQuestion, ColorMixQuestion, SequencesQuestion,
  GeographyQuestion, HumanBodyQuestion
- Delete IsraelQuestion and NatureQuestion thin wrappers (14 lines each)
  and inline CategoryIndexedQuestion with theme/colors directly in
  quizGameRegistry
