--- gamesformykids/components/game/quiz/screens/NatureQuestion.tsx
@@ -1,15 +1,0 @@
-'use client';
-import { CategoryIndexedQuestion } from '@/components/game/quiz';
-import type { NatureQuestion as Q } from '@/lib/quiz/data/nature';
-import { CATEGORY_COLORS } from '@/lib/quiz/data/nature';
-
-interface Props {
-  current: Q;
-  choices: string[];
-  correctLabel: string;
-  onSelect: (v: string) => void;
-}
-
-export default function NatureQuestion({ current, choices, correctLabel, onSelect }: Props) {
-  return <CategoryIndexedQuestion current={current} choices={choices} correctLabel={correctLabel} onSelect={onSelect} theme="green" categoryColors={CATEGORY_COLORS} />;
-}
